Reverted back to manual sprite-file creation. The spritesmith plugin expects you to override existing element styles with generated "icon-" classes, but this method fails if your elements are styled with IDs or if the source image changes with media queries (which is often the case with images on this site). I tested dynamically using "#name.png" or ".name.png" based on the original selectors (instead of using "icon-name.png"), but the problem still occurs.
Manually updating the css (in step 2) is less automated and more tedious, but it is the only feasible solution at the moment. Image changes on this site don't occur often, however, and this method has the added benefit of always knowing that the images will work. Here is the current build process:

If any images were added, removed, or modified, then
1. Run "grunt spritefile" to create a new spritefile ()
2. MANUALLY update styles.less file with the new sprite "background-position" css.
3. Run the default "grunt" task.

If no images changed:
1. Run the default "grunt" task.